+++ src/scenarios/client/http-custom-headers.ts
@@ -12,12 +12,12 @@
 
 import http from 'http';
 import {
-  Scenario,
   ScenarioUrls,
   ConformanceCheck,
   SpecVersion,
   DRAFT_PROTOCOL_VERSION
 } from '../../types.js';
+import { BaseHttpScenario } from './http-base.js';
 
 const SPEC_REFERENCE_CUSTOM = {
   id: 'SEP-2243-Custom-Headers',
@@ -131,139 +131,6 @@ function compareNumericValues(
     }
   }
   return null;
-}
-
-// Shared server boilerplate for Scenario implementations
-abstract class BaseHttpScenario implements Scenario {
-  abstract name: string;
-  abstract description: string;
-  abstract specVersions: SpecVersion[];
-  allowClientError?: boolean;
-
-  protected server: http.Server | null = null;
-  protected checks: ConformanceCheck[] = [];
-  protected port: number = 0;
-  protected sessionId: string = `session-${Date.now()}`;
-
-  async start(): Promise<ScenarioUrls> {
-    return new Promise((resolve, reject) => {
-      this.server = http.createServer((req, res) => {
-        this.handleRequest(req, res);
-      });
-      this.server.on('error', reject);
-      this.server.listen(0, () => {
-        const address = this.server!.address();
-        if (address && typeof address === 'object') {
-          this.port = address.port;
-          resolve({ serverUrl: `http://localhost:${this.port}` });
-        } else {
-          reject(new Error('Failed to get server address'));
-        }
-      });
-    });
-  }
-
-  async stop(): Promise<void> {
-    return new Promise((resolve, reject) => {
-      if (this.server) {
-        this.server.close((err) => {
-          if (err) reject(err);
-          else {
-            this.server = null;
-            resolve();
-          }
-        });
-      } else {
-        resolve();
-      }
-    });
-  }
-
-  abstract getChecks(): ConformanceCheck[];
-
-  protected handleRequest(
-    req: http.IncomingMessage,
-    res: http.ServerResponse
-  ): void {
-    if (req.method === 'GET') {
-      res.writeHead(200, {
-        'Content-Type': 'text/event-stream',
-        'Cache-Control': 'no-cache',
-        Connection: 'keep-alive',
-        'mcp-session-id': this.sessionId
-      });
-      res.write('data: \n\n');
-      return;
-    }
-    if (req.method === 'DELETE') {
-      res.writeHead(200);
-      res.end();
-      return;
-    }
-    if (req.method !== 'POST') {
-      res.writeHead(405);
-      res.end('Method Not Allowed');
-      return;
-    }
-
-    let body = '';
-    req.on('data', (chunk) => {
-      body += chunk.toString();
-    });
-    req.on('end', () => {
-      try {
-        const request = JSON.parse(body);
-        this.handlePost(req, res, request);
-      } catch (error) {
-        res.writeHead(400, { 'Content-Type': 'application/json' });
-        res.end(
-          JSON.stringify({
-            jsonrpc: '2.0',
-            error: { code: -32700, message: `Parse error: ${error}` }
-          })
-        );
-      }
-    });
-  }
-
-  protected abstract handlePost(
-    req: http.IncomingMessage,
-    res: http.ServerResponse,
-    request: any
-  ): void;
-
-  protected sendJson(res: http.ServerResponse, body: object): void {
-    res.writeHead(200, {
-      'Content-Type': 'application/json',
-      'mcp-session-id': this.sessionId
-    });
-    res.end(JSON.stringify(body));
-  }
-
-  protected sendInitialize(res: http.ServerResponse, request: any): void {
-    this.sendJson(res, {
-      jsonrpc: '2.0',
-      id: request.id,
-      result: {
-        protocolVersion: DRAFT_PROTOCOL_VERSION,
-        serverInfo: { name: this.name + '-server', version: '1.0.0' },
-        capabilities: { tools: {} }
-      }
-    });
-  }
-
-  protected sendNotificationAck(res: http.ServerResponse): void {
-    res.writeHead(202);
-    res.end();
-  }
-
-  protected sendGenericResult(res: http.ServerResponse, request: any): void {
-    this.sendJson(res, {
-      jsonrpc: '2.0',
-      id: request.id,
-      result: {}
-    });
-  }
 }
 
 // ─────────────────────────────────────────────────────────────────────────────
